+++ src/App.tsx
@@ -1,9 +1,10 @@
 import { ReactElement } from 'react'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
+import 'antd/dist/antd.css'
 import './App.css'
 import DataRegister from './pages/DataRegister/DataRegister'
 import Main from './pages/Main/Main'
-import MeasureRegister from './pages/MeasureRegister/MeasureForm'
+import MeasureRegister from './pages/MeasureRegister/MeasureRegister'
 
 const router = createBrowserRouter([
     {
